refactor: reorg ObjectiveTokenData.lua to Scenario/Marker/
