+++ TTS_xwing/src/Device/Remote/Commandos.lua
@@ -1,5 +1,5 @@
 local Dim = require("TTS_xwing.src.Dim")
-ObjectiveTokenData = require("TTS_xwing.src.ObjectiveTokenData")
+ObjectiveTokenData = require("TTS_xwing.src.Marker.Scenario.ObjectiveTokenData")
 
 ship_drop_positions = {
     small = {
